picochan-1.5
In this version, brotli compression support has been removed (hopefully only
temporarily) because the brotli module that was being used was unmaintained
and does not compile on OpenBSD despite considerable tinkering. Hopefully, a
new brotli module (perhaps using bindings to google's C library) can be found
or written.

This is not a huge deal because tor browser doesn't support brotli without
editing about:config. However, it is still intended that brotli support be
re-added in the future.

Other changes:
  - Small formatting engine fixes
  - Final database refinements