Fix string tokenizer
Previously, attempting to tokenize on "." would produce incorrect
results. This is because by using gmatch it is necessary to escape "."
like so: "%.". Because the delimiter is directly appended to the input
string, the last element in the resulting array has "%" appended to it.
This implementation disables regex altogether, but even when re-enabling
regex (by removing the last argument to string.find), correct results
are produced.