Add post tooltips when hovering over referrers/references
This change implements post previews at the expense of efficiency.
Basically, on the generation of each post, any post which the given post
references is also generated. Thus, post generation is duplicated. This
increases thread generation by a factor of the average amount of
references + referrers which are present in each post.