Combine multiple SQL statements into one.
The code for pico.file.list can be condensed by using SQL JOIN. The
resulting code doesn't need to be guarded by a transaction because all
single statements in SQL are automatically transactions. Likewise, the
code in pico.post.thread can be reduced similarly.

(This commit is courtesy of neeshy)