Always link to /<board>/<parent>#<number>.
On the recent page, invalid references would be generated if the
referenced post was in the same thread as the post making the reference
and the referenced post did not exist on the recent page. With this
change, references and referrers will always have valid links.

(This commit is courtesy of neeshy)