Correct calling of sqlite.open() and correctly print error message.
Previously, in picoengine.lua, an erroneous method was used to call
sqlite.open() and this resulted in unexpected behavior when the
database file did not exist. This change makes the error message
much clearer and also does not generate rubbish files if the user
has installed picochan and not created the database.