html.picofmt: move URL processing after other markup
This fixes a bug which would produce incorrect results when markup was
used within a URL. This is because the format processing is unable to
differentiate between text held within HTML attributes and text outside
of tags (the latter is the intended target for such processing). The
draw drawback of this solution is that URL processing is effectively
disabled for those cases. Also, b4c796a3 was reverted because it
interferes with this solution.