html.picofmt: simplify block code handling
With this an inline code block may appear at the beginning of a line,
whereas before, non-inline code blocks would take precendence in that
case. Incorrect results are (still) produced when following a non-inline
code block with a inline code block. The formatter produces a spurious
newline, however, this is preferable over the alternative.