html.picofmt: URL formatting should not take precedence
Previously, when an end delimeter was placed immediately following a
valid URL, the end delimeter would be parsed as though it were a part of
the URL. Conducting URL formatting after other formatting solved this,
but presented other, already known, problems. To avoid the prefix
pattern being "(.-)" (which causes slowdown), it is now "(.?.?.?.?.?.)".
This would have caused URLs placed at the beginning of the comment to be
parsed incorrectly, since the schema is variadic (it uses *). The prefix
pattern would take precedence, thus causing the following parse result:
  - string before URL parsing: "\nhttps://example.com/"
  - previous: "\nhttp"
  - url: "s://example.com/"
To satisfy both the prefix and schema pattern, six newlines are used to
prefix the comment as opposed to just one. Thus, there are always enough
characters preceding the URL for the prefix pattern to consume.