html.picofmt: revert to a narrower URL pattern
Using a variadic protocol pattern leads to performance issues because
Lua pattern matching is very unoptimized. A repetition before a single
character will cause O(n^2) time complexity whereas O(n) time complexity
is possible. The algorithm will naively search for the largest substring
of the input which matches the pattern until it fails, and then start
again with the next substring. An optimized version of the algorithm
would search for the single character and use it as an anchor (similar
to ^ or $). Then it would search backwards and forwards to complete the
match. For the worst case in the former, the same character will be
scanned n^2, where n is the length of the input. For the latter, no
character will be scanned more than twice (especially because matches
cannot overlap). Other solutions are possible. Assuming that a character
class in the pattern is mutually exclusive with the previous (which is a
repetition), a table an be used to "remember" where the repetition ends.
Thus, if the substring search fails, it can be skipped for the next
iteration.

Manually programming a matching algorithm for this specific case may be
performant. Alternatively, writing FFI bindings for regex.h would be
another solution since regular expression engines don't suffer from this
problem.