html.renderpostfiles: use img instead of div for post files
An img tag with the lazy loading feature has the same effect as a
hidden-by-default div with a background image. Additionally, whereas
before the height and width needed to be explicitly specified, now the
img tag carries that information by default, so the max-width property
can be independently set to 100%.

This is now much better. Before this change, expanding a small image
would unconditionally fill the entire screen.

Finally, the floating right property has been removed.