Webring: remove
Webring support is half baked, non-performant and a security risk.
In other webring implementations, much of the work involved in
displaying remote boards is offloaded onto the client through
javascript. Since picochan does not make use of javascript, curl on the
server side must be used instead. Binding to and calling into libcurl
introduces a massive attack surface, and produces performance penalties
when accessing the /Boards page. The asynchronous curl multi-API could
be used be used instead, but that is still too prohibitive for an
otherwise snappy imageboard engine.